+++ drizzle/migrations/0004_mysterious_misty_knight.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `Settings` ADD `enableFileWatcher` integer DEFAULT true NOT NULL;